+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 Changelog
 =========
 
+* **2013-12-27**: Added XmlSchemaTestCase to test XML schema's
 * **2013-11-17**: Added DatabaseTestListener to support database testing